+++ sperl_vm.c
@@ -1168,7 +1168,7 @@ void SPerl_VM_call_sub(SPerl* sperl, SPerl_VM* vm, const char* sub_base_name) {
         // Update call stack base
         call_stack_base = call_stack_next;
         
-        operand_stack_top -= sub->op_args->length;
+        operand_stack_top -= (int32_t) sub->op_args->length;
         
         CALLSUB_COMMON:
